requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
anthropic==0.49.0
httpx==0.28.1
pydantic==2.10.6
mcp==1.3.0
pgvector==0.3.6
